move config package under cli
